Problem 10 sum of primes below 2,000,000 BLUE use Eratosthenes sieve

+++ net.projecteuler.solutions/src/test/java/problem/no1to10/Problem10Test.java
@@ -22,6 +22,6 @@ class Problem10Test {
     }
 
     private static void assertEquals(int expected, int i) {
-        Assertions.assertEquals(BigInteger.valueOf(expected), Problem10.sumOfPrimesBelow(i));
+        Assertions.assertEquals(expected, Problem10.sumOfPrimesBelow(i));
     }
 }
